# openapi.yaml
openapi: 3.0.1
info:
  title: File Upload / Download API
  version: 1.0.0
  description: API for file uploads and downloads
tags:
  - name: file
    description: API for managing files
paths:
  /:
    get:
      tags:
        - file
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileMetaInfoList'
          description: listed files
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Generic error
      operationId: listFiles
      summary: List all files.
  /upload-file:
    post:
      tags:
        - file
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                attachment:
                  format: binary
                  description: file to be uploaded
                  type: string
            encoding:
              attachment:
                contentType: application/octet-stream
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileMetaInfo'
          description: uploaded file
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Generic error
      operationId: uploadFile
      summary: Uploads a file.
  /upload-files:
    post:
      tags:
        - file
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                attachments:
                  description: files to be uploaded
                  type: array
                  items:
                    format: binary
                    type: string
            encoding:
              attachments:
                contentType: application/octet-stream
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileMetaInfoList'
          description: uploaded files
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Generic error
      operationId: uploadFiles
      summary: Uploads multiple files.
  /download/{fileName}:
    get:
      tags:
        - file
      summary: Downloads a file.
      operationId: downloadFile
      parameters:
        - in: path
          name: fileName
          schema:
            type: string
          required: true
          description: file name
      responses:
        '200':
          description: file
          content:
            application/octet-stream:
              schema:
                type: string
                format: byte
                description: resource as bytes
        '500':
          description: Generic error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
components:
  schemas:
    FileMetaInfoList:
      description: list of FileMetaInfo objects
      type: array
      items:
        $ref: '#/components/schemas/FileMetaInfo'
    FileMetaInfo:
      description: response object for uploaded file
      type: object
      properties:
        fileName:
          nullable: false
          description: file name
          type: string
          example: test.pdf
        downloadUrl:
          nullable: false
          description: download url
          type: string
          example: 'http://localhost/download/test.pdf'
        contentType:
          nullable: false
          type: string
          example: application/pdf
        size:
          format: int64
          description: file size in kb
          type: integer
          example: 5
    ApiError:
      required:
        - errorId
        - errorMessage
      type: object
      properties:
        errorId:
          description: error id
          type: string
          example: 4711
        errorMessage:
          description: error message
          type: string
          example: Object not found